Pigean cleanup (#797)
* gene page moved stuff to module

* refactored all pigean pages to use utils

* label fix
