+++ src/views/PIGEAN/NetworkGraph/main.js
@@ -6,6 +6,7 @@ import NetworkGraph from "@/components/NetworkGraph.vue";
 import SearchHeaderWrapper from "@/components/SearchHeaderWrapper.vue";
 import GenesetSizeSelectPicker from "@/components/GenesetSizeSelectPicker.vue";
 import keyParams from "@/utils/keyParams";
+import pigeanUtils from "@/utils/pigeanUtils.js";
 import { pageMixin } from "@/mixins/pageMixin.js";
 new Vue({
     store,
@@ -35,7 +36,7 @@ new Vue({
     },
     methods: {
         setSelectedPhenotype(PHENOTYPE) {
-            let oldStylePhenotype = this.toOldStyle(PHENOTYPE);
+            let oldStylePhenotype = pigeanUtils.toOldStyle(PHENOTYPE);
             this.newPhenotypeSearchKey = oldStylePhenotype.description;
             this.phenotypeSearchKey = null;
             this.selectedPhenotype = oldStylePhenotype;
@@ -52,25 +53,10 @@ new Vue({
 
             return isInPhenotype == searchKeys.length ? true : null;
         },
-        toOldStyle(newStylePhenotype){
-            let oldStyle = structuredClone(newStylePhenotype);
-            oldStyle.description = newStylePhenotype.phenotype_name;
-            oldStyle.name = newStylePhenotype.phenotype;
-            oldStyle.group = newStylePhenotype.display_group;
-            return oldStyle;
-        },
         searchPhenotype(){
             let phenotypeToSearch = this.selectedPhenotype 
                 || this.pigeanPhenotypeMap[keyParams.phenotype];
             this.$store.dispatch("sendSearch", phenotypeToSearch);
-        },
-        mapPhenotypes(){
-            let phenotypeMap = {};
-            let phenotypes = this.$store.state.pigeanAllPhenotypes.data
-            phenotypes.forEach(item => {
-                phenotypeMap[item.phenotype] = this.toOldStyle(item);
-            });
-            return phenotypeMap;
         },
     },
     async created() {
@@ -78,7 +64,8 @@ new Vue({
         this.$store.dispatch("bioPortal/getDiseaseGroups");
         this.$store.dispatch("bioPortal/getPhenotypes");
         await this.$store.dispatch("getPigeanPhenotypes");
-        this.pigeanPhenotypeMap = this.mapPhenotypes();
+        this.pigeanPhenotypeMap = 
+            pigeanUtils.mapPhenotypes(this.$store.state.pigeanAllPhenotypes.data);
         let initialPhenotype = this.pigeanPhenotypeMap[keyParams.phenotype];
         this.$store.dispatch("sendSearch", initialPhenotype)
     },
